pyproject.toml
[tool.poetry]
name = "sscomp"
version = "0.1.0"
description = "Tooling for compiling jinja templates into a static web site"
authors = ["Ben Airey <benjaminjohnairey@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Jinja2 = "^3.1.3"
click = "^8.1.7"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pylint = "^3.1.0"
black = "^24.2.0"
bandit = "^1.7.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
